mergeObservables + a helper method for observing multiple observables
