+++ src/libs/observable-api/utils/index.ts
@@ -1,9 +1,10 @@
 export * from './append-to-frozen-array.function.js';
-export * from './combine-observables.function.js';
 export * from './create-observable-part.function.js';
 export * from './deep-freeze.function.js';
 export * from './default-memoization.function.js';
 export * from './filter-frozen-array.function.js';
+export * from './merge-observables.function.js';
 export * from './naive-object-comparison.function.js';
+export * from './observe-multiple.function.js';
 export * from './partial-update-frozen-array.function.js';
 export * from './push-to-unique-array.function.js';
